move utils into algo folder
